fix: remove query files
